+++ src/main/scala/edu/ie3/simona/model/participant2/PvModel.scala
@@ -26,14 +26,17 @@ import edu.ie3.simona.model.participant2.ParticipantModel.{
 import edu.ie3.simona.model.participant2.PvModel._
 import edu.ie3.simona.ontology.messages.services.WeatherMessage.WeatherData
 import edu.ie3.simona.service.ServiceType
-import edu.ie3.util.quantities.PowerSystemUnits
 import edu.ie3.util.quantities.QuantityUtils.RichQuantityDouble
 import edu.ie3.util.scala.quantities.DefaultQuantities.zeroWPerSM
+import edu.ie3.util.scala.quantities.QuantityConversionUtils.{
+  DimensionlessToSimona,
+  PowerConversionSimona,
+  RadiansConversionSimona,
+}
 import edu.ie3.util.scala.quantities._
 import squants._
 import squants.space.{Degrees, SquareMeters}
 import squants.time.Minutes
-import tech.units.indriya.unit.Units._
 
 import java.time.ZonedDateTime
 import java.util.UUID
@@ -271,35 +274,15 @@ object PvModel {
     new PvModel(
       input.getUuid,
       input.getId,
-      Kilovoltamperes(
-        input.getsRated
-          .to(PowerSystemUnits.KILOVOLTAMPERE)
-          .getValue
-          .doubleValue
-      ),
+      input.getsRated.toKilovoltamperes,
       input.getCosPhiRated,
       QControl(input.getqCharacteristics),
       Degrees(input.getNode.getGeoPosition.getY),
       Degrees(input.getNode.getGeoPosition.getX),
       input.getAlbedo,
-      Each(
-        input.getEtaConv
-          .to(PowerSystemUnits.PU)
-          .getValue
-          .doubleValue
-      ),
-      Radians(
-        input.getAzimuth
-          .to(RADIAN)
-          .getValue
-          .doubleValue
-      ),
-      Radians(
-        input.getElevationAngle
-          .to(RADIAN)
-          .getValue
-          .doubleValue
-      ),
+      input.getEtaConv.toPercent,
+      input.getAzimuth.toRadians,
+      input.getElevationAngle.toRadians,
     )
 
   /** Calculates the position of the earth in relation to the sun (day angle)